+++ photolist.h
@@ -10,7 +10,7 @@
     /* prototype for photo list node, contains all information regarding each 
      * photo as well as extras for replication purposes */
 typedef struct photo_node_t{
-    uint64_t photo_id;
+    uint32_t photo_id;
     char photo_name[50];
     int deleted;
     struct photo_node_t *prev;
@@ -22,6 +22,6 @@ typedef struct photo_node_t{
 photolist_t *photolist_init();
 /* add photo to photolist, to be inserted in sorted order according to photo_id */
 int photolist_insert(photolist_t **photos, uint32_t photo_id, char *photo_name, pthread_rwlock_t *rwlock);
-int photolist_delete(photolist_t **photos, uint64_t photo_id, pthread_rwlock_t *rwlock);
+int photolist_delete(photolist_t **photos, uint32_t photo_id, pthread_rwlock_t *rwlock);
 int photolist_print(photolist_t **photos, pthread_rwlock_t *rwlock);
 #endif //PHOTOLIST_H
